pyproject.toml
[tool.poetry]
name = "kargolast"
version = "0.1.0"
description = ""
authors = ["artykovv <pipzdty@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.111.0"}
sqlalchemy = "^2.0.30"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
greenlet = "^3.0.3"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"
openpyxl = "^3.1.4"
aiofiles = "^23.2.1"
jinja2 = "^3.1.4"
passlib = "^1.7.4"
xlsxwriter = "^3.2.0"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
pytz = "^2024.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
